update readme and schema file

+++ db/migrate/20231109091358_create_cars.rb
@@ -5,7 +5,7 @@ class CreateCars < ActiveRecord::Migration[7.1]
       t.string :model
       t.string :description
       t.string :image
-      t.decimal :price_per_day
+      t.integer :price_per_day
       t.integer :seating_capacity
 
       t.timestamps
